Move db state conversion with db type definition
Those conversion are specific to the database and should be defined with
in the database module.
